vst2: Implement a VST2.x compatible Denoiser effect
This makes VoiceFX available for not just VST3.x but also VST2.x hosts, opening many more doors for future work. While not all VST2.x hosts will work due to the ideal requirement of SDK 2.3 or later, this does cover the majority of big name DAWs and NLEs.

The functionality is mostly the same as the VST3.x version, just much more modernized and adjusted to work with the VST2.x API.